refactor: move SecretScanner and its UI classes to the `secret` package
