+++ ggshield/cmd/secret/scan/docker.py
@@ -10,7 +10,7 @@ from ggshield.cmd.secret.scan.secret_scan_common_options import (
 )
 from ggshield.core.errors import handle_exception
 from ggshield.scan import ScanContext, ScanMode
-from ggshield.scan.docker import docker_save_to_tmp, docker_scan_archive
+from ggshield.secret.docker import docker_save_to_tmp, docker_scan_archive
 
 
 # bailout if docker command takes longer than 6 minutes
