+++ ggshield/cmd/secret/scan/dockerarchive.py
@@ -9,7 +9,7 @@ from ggshield.cmd.secret.scan.secret_scan_common_options import (
 )
 from ggshield.core.errors import handle_exception
 from ggshield.scan import ScanContext, ScanMode
-from ggshield.scan.docker import docker_scan_archive
+from ggshield.secret.docker import docker_scan_archive
 
 
 @click.command(hidden=True)
