+++ ggshield/cmd/secret/scan/precommit.py
@@ -7,7 +7,8 @@ from ggshield.cmd.secret.scan.secret_scan_common_options import (
 )
 from ggshield.core.errors import handle_exception
 from ggshield.core.git_shell import check_git_dir
-from ggshield.scan import Commit, ScanCollection, ScanContext, ScanMode, SecretScanner
+from ggshield.scan import Commit, ScanCollection, ScanContext, ScanMode
+from ggshield.secret import SecretScanner
 from ggshield.secret.output import SecretTextOutputHandler
 from ggshield.secret.output.messages import remediation_message
