+++ ggshield/cmd/secret/scan/prepush.py
@@ -19,8 +19,8 @@ from ggshield.core.git_shell import (
 from ggshield.core.text_utils import display_warning
 from ggshield.core.utils import EMPTY_SHA, EMPTY_TREE
 from ggshield.scan import ScanContext, ScanMode
-from ggshield.scan.repo import scan_commit_range
 from ggshield.secret.output.messages import remediation_message
+from ggshield.secret.repo import scan_commit_range
 
 
 logger = logging.getLogger(__name__)
