+++ ggshield/scan/__init__.py
@@ -1,10 +1,9 @@
 from .commit import Commit
 from .file import File, get_files_from_paths
-from .rich_secret_scanner_ui import RichSecretScannerUI
 from .scan_context import ScanContext
 from .scan_mode import ScanMode
 from .scannable import DecodeError, Files, Scannable, StringScannable
-from .scanner import Result, Results, ScanCollection, SecretScanner, SecretScannerUI
+from .scanner import Result, Results, ScanCollection
 
 
 __all__ = [
@@ -15,9 +14,6 @@ __all__ = [
     "Files",
     "Result",
     "Results",
-    "RichSecretScannerUI",
-    "SecretScanner",
-    "SecretScannerUI",
     "ScanCollection",
     "ScanContext",
     "ScanMode",
