+++ ggshield/secret/__init__.py
@@ -1,0 +1,9 @@
+from .rich_secret_scanner_ui import RichSecretScannerUI
+from .secret_scanner import SecretScanner, SecretScannerUI
+
+
+__all__ = [
+    "RichSecretScannerUI",
+    "SecretScanner",
+    "SecretScannerUI",
+]
