+++ tests/unit/cmd/scan/test_docker.py
@@ -8,7 +8,7 @@ import pytest
 from ggshield.cmd.main import cli
 from ggshield.core.errors import ExitCode
 from ggshield.scan import Files, ScanCollection, StringScannable
-from ggshield.scan.docker import DockerImage, LayerInfo, _validate_filepath
+from ggshield.secret.docker import DockerImage, LayerInfo, _validate_filepath
 from tests.unit.conftest import (
     DOCKER__INCOMPLETE_MANIFEST_EXAMPLE_PATH,
     DOCKER_EXAMPLE_PATH,
@@ -94,7 +94,7 @@ class TestDockerCMD:
         assert_invoke_exited_with(result, ExitCode.UNEXPECTED_ERROR)
         assert 'Image "ggshield-non-existant" not found' in result.output
 
-    @patch("ggshield.scan.docker.DockerImage")
+    @patch("ggshield.secret.docker.DockerImage")
     @pytest.mark.parametrize(
         "image_path", [DOCKER_EXAMPLE_PATH, DOCKER__INCOMPLETE_MANIFEST_EXAMPLE_PATH]
     )
