+++ tests/unit/secret/test_scan.py
@@ -6,9 +6,10 @@ from click import Command, Context, Group
 
 from ggshield import __version__
 from ggshield.core.cache import Cache
-from ggshield.scan import Commit, ScanContext, ScanMode, SecretScanner
-from ggshield.scan.repo import cd
+from ggshield.scan import Commit, ScanContext, ScanMode
 from ggshield.scan.scan_context import get_os_info
+from ggshield.secret import SecretScanner
+from ggshield.secret.repo import cd
 from tests.unit.conftest import UNCHECKED_SECRET_PATCH
