+++ tests/unit/secret/test_scan_docset.py
