+++ tests/unit/secret/test_scan_repo.py
@@ -6,7 +6,7 @@ import pytest
 
 from ggshield.scan import Commit, Result, Results, StringScannable
 from ggshield.scan.commit import CommitInformation
-from ggshield.scan.repo import get_commits_by_batch, scan_commits_content
+from ggshield.secret.repo import get_commits_by_batch, scan_commits_content
 from tests.unit.conftest import TWO_POLICY_BREAKS
 
 
@@ -70,7 +70,7 @@ def test_get_commits_content_by_batch(
         assert len(batch) == len(expected_batch)
 
 
-@patch("ggshield.scan.repo.SecretScanner")
+@patch("ggshield.secret.repo.SecretScanner")
 def test_scan_2_commits_same_content(secret_scanner_mock):
     """
     GIVEN 2 commits where each commit has a file with same content and same filename
@@ -117,7 +117,7 @@ def test_scan_2_commits_same_content(secret_scanner_mock):
     assert all_policy_breaks_count == 4
 
 
-@patch("ggshield.scan.repo.SecretScanner")
+@patch("ggshield.secret.repo.SecretScanner")
 def test_scan_2_commits_file_association(secret_scanner_mock):
     """
     GIVEN 2 commits with several files in some commits
